+++ src/util/test_system_directory_custom_name.rs
@@ -31,6 +31,8 @@ async fn test() {
         completion: 0,
         merging: Merging::NonMerged as i64,
         custom_extension: None,
+        games_complete: 0,
+        games_total: 0,
     };
 
     // when